feat: enhance realtime trigger lifecycle
part-of: kestra-io/kestra#3767
