+++ src/main/java/io/kestra/plugin/aws/sqs/RealtimeTrigger.java
@@ -16,8 +16,13 @@ import lombok.*;
 import lombok.experimental.SuperBuilder;
 import org.reactivestreams.Publisher;
 import reactor.core.publisher.Flux;
+import software.amazon.awssdk.services.sqs.SqsAsyncClient;
+import software.amazon.awssdk.services.sqs.model.DeleteMessageRequest;
+import software.amazon.awssdk.services.sqs.model.ReceiveMessageRequest;
 
 import java.time.Duration;
+import java.util.concurrent.CountDownLatch;
+import java.util.concurrent.atomic.AtomicBoolean;
 
 @SuperBuilder
 @ToString
@@ -68,6 +73,14 @@ public class RealtimeTrigger extends AbstractTrigger implements RealtimeTriggerI
     @Builder.Default
     protected Duration stsRoleSessionDuration = AbstractConnectionInterface.AWS_MIN_STS_ROLE_SESSION_DURATION;
 
+    @Builder.Default
+    @Getter(AccessLevel.NONE)
+    private final AtomicBoolean isActive = new AtomicBoolean(true);
+
+    @Builder.Default
+    @Getter(AccessLevel.NONE)
+    private final CountDownLatch waitForTermination = new CountDownLatch(1);
+
     @Override
     public Publisher<Execution> evaluate(ConditionContext conditionContext, TriggerContext context) throws Exception {
         RunContext runContext = conditionContext.getRunContext();
@@ -87,7 +100,81 @@ public class RealtimeTrigger extends AbstractTrigger implements RealtimeTriggerI
             .stsEndpointOverride(this.stsEndpointOverride)
             .build();
 
-        return Flux.from(task.stream(conditionContext.getRunContext()))
+        return Flux.from(publisher(task, conditionContext.getRunContext()))
             .map(record -> TriggerService.generateRealtimeExecution(this, context, record));
+    }
+
+    public Flux<Message> publisher(final Consume task,
+                                   final RunContext runContext) throws Exception {
+        var queueUrl = runContext.render(getQueueUrl());
+
+        return Flux.create(
+            fluxSink -> {
+                try (SqsAsyncClient sqsClient = task.asyncClient(runContext)) {
+                    while (isActive.get()) {
+                        ReceiveMessageRequest receiveRequest = ReceiveMessageRequest.builder()
+                            .queueUrl(queueUrl)
+                            .build();
+
+                        sqsClient.receiveMessage(receiveRequest)
+                            .whenComplete((messageResponse, throwable) -> {
+                                if (throwable != null) {
+                                    fluxSink.error(throwable);
+                                } else {
+                                    messageResponse.messages().forEach(message -> {
+                                        fluxSink.next(Message.builder().data(message.body()).build());
+                                    });
+                                    messageResponse.messages().forEach(message ->
+                                        sqsClient.deleteMessage(DeleteMessageRequest.builder()
+                                            .queueUrl(queueUrl)
+                                            .receiptHandle(message.receiptHandle())
+                                            .build()
+                                        )
+                                    );
+                                }
+                            });
+                        try {
+                            Thread.sleep(100);
+                        } catch (InterruptedException e) {
+                            Thread.currentThread().interrupt();
+                            isActive.set(false); // proactively stop polling
+                        }
+                    }
+                } catch (Throwable e) {
+                    fluxSink.error(e);
+                } finally {
+                    fluxSink.complete();
+                    this.waitForTermination.countDown();
+                }
+            });
+    }
+
+    /**
+     * {@inheritDoc}
+     **/
+    @Override
+    public void kill() {
+        stop(true);
+    }
+
+    /**
+     * {@inheritDoc}
+     **/
+    @Override
+    public void stop() {
+        stop(false); // must be non-blocking
+    }
+
+    private void stop(boolean wait) {
+        if (!isActive.compareAndSet(true, false)) {
+            return;
+        }
+        if (wait) {
+            try {
+                this.waitForTermination.await();
+            } catch (InterruptedException e) {
+                Thread.currentThread().interrupt();
+            }
+        }
     }
 }
